# internal/modules/lab/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.0

package dbgen

type Event struct {
	ID          string
	RunID       string
	Seq         int64
	Kind        string
	PayloadJson string
	CreatedAt   string
}

type Run struct {
	ID              string
	Title           string
	ScenarioID      string
	Mode            string
	Status          string
	ProtocolVersion string
	AgentType       string
	CreatedAt       string
	FinishedAt      string
	LastSeq         int64
}